validaciones

+++ empleados/index.php
@@ -11,6 +11,9 @@ $txtFoto = (isset($_FILES['txtFoto']["name"]))?$_FILES['txtFoto']["name"]:"";
 
 $accion = (isset($_POST['accion']))?$_POST['accion']:"";
 
+$error = array();
+
+/* activacion de botones */
 $accionAgregar = "";
 $accionModificar = $accionEliminar = $accionCancelar = "disabled";
 $mostrarModal = false;
@@ -20,6 +23,24 @@ include("../conexion/conexion.php");
 
 switch ($accion) {
     case 'btnAgregar':
+
+        /* validacion */
+        if ($txtNombre=="") {
+            $error['nombre'] = "Escribe el nombre";
+        }
+        if ($txtApellido=="") {
+            $error['apellido'] = "Escribe el apellido";
+        }
+        if ($txtCorreo=="") {
+            $error['correo'] = "Correo no valido";
+        }
+        if ($txtTelefono=="") {
+            $error['telefono'] = "Digita tu numero de contacto";
+        }
+        if (count($error) > 0) {
+            $mostrarModal = true;
+        break;
+        }
 
         $insertar = $pdo->prepare("INSERT INTO empleados(nombre,apellido,correo,telefono,foto) /* Creamos un objeto  */
         VALUES(:nombre,:apellido,:correo,:telefono,:foto)");
@@ -42,8 +63,6 @@ switch ($accion) {
 
         header('Location: index.php');
 
-        echo $txtid;
-        echo "presionaste agregar";
         break;
 
     case 'btnModificar':
@@ -78,7 +97,10 @@ switch ($accion) {
 
         if (isset($empleado["foto"])) {
             if (file_exists("../imagenes/".$empleado["foto"])) {
-                unlink("../imagenes/".$empleado["foto"]);
+
+                if ($item['foto']!="cheems.png") {
+                    unlink("../imagenes/".$empleado["foto"]);
+                }
             }
         }
 
@@ -92,8 +114,7 @@ switch ($accion) {
         
 
         header('Location: index.php');
-        echo $txtid;
-        echo "presionaste modificar";
+        
         break;
     case 'btnEliminar':
 
@@ -104,7 +125,7 @@ switch ($accion) {
         $empleado = $insertar->fetch(PDO::FETCH_LAZY);
         print_r($empleado);
 
-        if (isset($empleado['foto'])) {
+        if (isset($empleado['foto'])&&($item['foto']!="cheems.png")) {
             if (file_exists('../imagenes/'.$empleado['foto'])) {
                 unlink("../imagenes/".$empleado["foto"]);
             }
@@ -116,11 +137,29 @@ switch ($accion) {
         $insertar->execute();
 
         header('Location: index.php');
-        echo $txtid;
-        echo "presionaste Eliminar";
+       
         break;
     case 'btnCancelar':
-        # code...
+        header('Location: index.php');
+        break;
+
+    case 'Seleccionar':
+        $accionAgregar = "disabled";
+        $accionModificar = $accionEliminar = $accionCancelar = "";
+        $mostrarModal=true;
+
+        $insertar = $pdo->prepare("SELECT * FROM empleados WHERE id=:id"); 
+        $insertar->bindParam(':id',$txtid);
+        $insertar->execute();
+        $empleado = $insertar->fetch(PDO::FETCH_LAZY);
+
+        $txtNombre = $empleado['nombre'];
+        $txtApellido = $empleado['apellido'];
+        $txtCorreo = $empleado['correo'];
+        $txtTelefono = $empleado['telefono'];
+        $txtFoto = $empleado['foto'];
+
+
         break;
     
     default:
@@ -175,30 +214,49 @@ switch ($accion) {
 
                 <div class="form-group col-md-6">
                 <label for="">Nombre(S):</label>
-                <input type="text" class="form-control" name="txtNombre" value="<?php echo $txtNombre; ?>" placeholder="" id="txtNombre" required="">
+                <input type="text" class="form-control <?php echo(isset($error['nombre']))?"is-invalid":""; ?>" name="txtNombre" value="<?php echo $txtNombre; ?>" placeholder="" id="txtNombre" required="">
+                <div class="invalid-feedback">
+                <?php echo(isset($error['nombre']))?$error['nombre']:""; ?>
+                </div>
                 <br>
                 </div>
                 
                 <div class="form-group col-md-6">
                 <label for="">Apellido(s):</label>
-                <input type="text" class="form-control" name="txtApellido" value="<?php echo $txtApellido; ?>" placeholder="" id="txt3" required="">
+                <input type="text" class="form-control <?php echo(isset($error['apellido']))?"is-invalid":""; ?>" name="txtApellido" value="<?php echo $txtApellido; ?>" placeholder="" id="txt3" required="">
+                <div class="invalid-feedback">
+                <?php echo(isset($error['apellido']))?$error['apellido']:""; ?>
+                </div>
                 <br>
                 </div>
 
                 <div class="form-group col-md-12">
                 <label for="">Correo:</label>
-                <input type="email" class="form-control" name="txtCorreo" value="<?php echo $txtCorreo; ?>" placeholder="" id="txt4" required="">
+                <input type="email" class="form-control <?php echo(isset($error['correo']))?"is-invalid":""; ?>" name="txtCorreo" value="<?php echo $txtCorreo; ?>" placeholder="" id="txt4" required="">
+                <div class="invalid-feedback">
+                <?php echo(isset($error['correo']))?$error['correo']:""; ?>
+                </div>
                 <br>
                 </div>
 
                 <div class="form-group col-md-4">
                 <label for="">Telefono:</label>
-                <input type="text" class="form-control" name="txtTelefono" value="<?php echo $txtTelefono; ?>" placeholder="" id="txt5" required="">
+                <input type="text" class="form-control <?php echo(isset($error['telefono']))?"is-invalid":""; ?>" name="txtTelefono" value="<?php echo $txtTelefono; ?>" placeholder="" id="txt5" required="">
+                <div class="invalid-feedback">
+                <?php echo(isset($error['telefono']))?$error['telefono']:""; ?>
+                </div>
                 <br>
                 </div>
 
                 <div class="form-group col-md-12">
                 <label for="">Foto:</label>
+                <?php if ($txtFoto!="") { ?>
+                    </br>
+                        <img src="../imagenes/<?php echo $txtFoto; ?>" alt="" class="img-thumbnail rounded mx-auto d-block" width="100px">
+                    </br>
+                    </br>
+                <?php } ?>
+
                 <input type="file" class="form-control" accept="image/*" value="<?php echo $txtFoto; ?>" name="txtFoto" placeholder="" id="txt6" >
                 <br>
                 </div>
@@ -206,15 +264,17 @@ switch ($accion) {
                 </div>
             </div>
             <div class="modal-footer">
-                <button value="btnAgregar" class="btn btn-success" type="submit" name="accion">Agregar</button>
-                <button value="btnModificar" class="btn btn-warning" type="submit" name="accion">Modificar</button>
-                <button value="btnEliminar" class="btn btn-danger" type="submit" name="accion">Eliminar</button>
-                <button value="btnCancelar" class="btn btn-primary" type="submit" name="accion">Cancelar</button>
+                <button value="btnAgregar" <?php echo $accionAgregar ?> class="btn btn-success" type="submit" name="accion">Agregar</button>
+                <button value="btnModificar" <?php echo $accionModificar ?> class="btn btn-warning" type="submit" name="accion">Modificar</button>
+                <button value="btnEliminar" <?php echo $accionEliminar ?> class="btn btn-danger" type="submit" name="accion">Eliminar</button>
+                <button value="btnCancelar" <?php echo $accionCancelar ?> class="btn btn-primary" type="submit" name="accion">Cancelar</button>
 
             </div>
         </div>
         </div>
 </div>
+<br>
+<br>
         <!-- Button trigger modal -->
         <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#exampleModal">
             Agregar
@@ -247,16 +307,13 @@ switch ($accion) {
                         <!-- cuando seleccionemos se enviaran los datos -->
                         <td>
                         <form action="" method="post" ectype="multipart/form-data">
-                        <input type="hidden" name="txtid" value="<?php echo $empleado['id']; ?>">
-                        <input type="hidden" name="txtNombre" value="<?php echo $empleado['nombre']; ?>">
-                        <input type="hidden" name="txtApellido" value="<?php echo $empleado['apellido']; ?>">
-                        <input type="hidden" name="txtCorreo" value="<?php echo $empleado['correo']; ?>">
-                        <input type="hidden" name="txtTelefono" value="<?php echo $empleado['telefono']; ?>">
-                        <input type="hidden" name="txtFoto" value="<?php echo $empleado['foto']; ?>">
 
-                        <input type="submit" value="Seleccionar" name="accion">
-                        <button value="btnEliminar" type="submit" name="accion">Eliminar</button>
-                        
+                        <input type="hidden" name="txtid" value="<?php echo $empleado['id']; ?>">
+
+                        <input type="submit" value="Seleccionar" class="btn btn-info" name="accion">
+                        <button value="btnEliminar" type="submit" class="btn btn-danger" name="accion">Eliminar</button>
+
+             
                         
                         </form>
                         
@@ -266,7 +323,12 @@ switch ($accion) {
             
                <?php } ?>
             </table>
+            
+                    
         </div>
+
+
+        
     </div>
 
 
@@ -278,6 +340,9 @@ switch ($accion) {
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.min.js" integrity="sha384-w1Q4orYjBQndcko6MimVbzY0tgp4pWB4lZ7lr30WKz0vr/aWKhXdBNmNb5D92v7s" crossorigin="anonymous"></script>
-    
+<?php if ($mostrarModal) {?>
+            <script>$('#exampleModal').modal('show');</script>
+                      
+<?php } ?> 
 </body>
 </html>
